feat: match on pkgAttrName (#63)
Modifies `getPackages` and `PkgDbQuery` to use `match` argument to match a calculated column `pkgAttrName`.

For `catalog` subtrees this is the _second to last member of an attribute path_, e.g. `["catalog", "x86_64-linux", "stable", "hello", "latest"] => "hello"`.
For `packages` and `legacyPackages` subtrees this is the _last member of an attribute path_, e.g. `["catalog", "x86_64-linux", "hello"] => "hello"`.

Additionally this fixes a bug in the error output of `QueryPreferences` parsers which caused incorrect attribute paths like `allow.allow` and `semver.semver` to be emitted instead of `allow.bogus` and `semver.bogus`.