fix: link bear wrapper to bear.d and add to path
bear uses two strategies to build its compile database.

1. by injecting a compiler library wrapper via LD_PRELOAD
2. by using a compiler wrapper binary

On macos only the latter is supported.
However, bear just sets the `CXX` variable to `c++`, seemingly assuming that `c++` on the PATH is their wrapper.
My explanation for that is that bear also sets `PATH` to include `<bear>/lib/bear/wrapper.d`,
which usually contains links to the wrapper but is not populated by the nix built bear.

Thus the solution here is to create a link ourselves and add that to the PATH.
I am not sure how well that works on linux or if we should make this conditional
(linux likely uses the LD_PRELOAD method by default).