feat(lock): reuse entries from old lockfile (#130)
Rework `lockSystem` to operate over locked groups (calculated by
`getLockedGroups`) and unlocked groups (calculated by
`getUnlockedGroups`), instead of descriptors and grouped descriptors.

Unlocked groups are re-resolved, whereas existing entries from the lock
are copied for locked groups.

All descriptors with no group are added to a single group which is
resolved together; this behavior could be configured differently in the
future.

`fillLockedFromOldLockfile` and `getUngroupedDescriptors` are dropped as
they are unused after these changes.

---------

Co-authored-by: Alex Ameen <alex.ameen.tx@gmail.com>