feat: parse descriptors from strings (#195)
This adds a `pkgdb parse descriptor` subcommand that parses descriptors
provided as strings in the format:
```
[<input>:]((<attr>.)*<attrName>|<pname>)[@(<semver>|=<version>)]
```

The output format is always JSON, but there is an optional `--to` flag
which determines which representation of the descriptor is dumped as
JSON. Available options are `manifest` (the default) and `query`.

The `manifest` option dumps the JSON representation of the descriptor as
if it had been specified in the manifest. The `query` option dumps the
JSON representation of the descriptor after it has been transformed into
a query. This second option is useful for verifying that a descriptor
has been converted to the intended query (e.g. that the version
requirement has populated `semver` or `version` as intended).

---------

Co-authored-by: Alex Ameen <alex.ameen.tx@gmail.com>